requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pynacl==1.5.0
python-dotenv
flask>=3.0.0
cryptography>=42.0.0
python-dotenv>=1.0.0  # для .env-файла с API_KEY